+++ src/Datasets/Generators/HalfMoon.php
@@ -115,7 +115,7 @@ class HalfMoon implements Generator
             ->add($this->center)
             ->asArray();
 
-        return new Unlabeled($samples, false);
+        return Unlabeled::quick($samples);
     }
 
 }