+++ CHANGELOG.md
@@ -17,6 +17,7 @@
     - Renamed restore method to load on Datasets and Persisters
     - Random Forest now accepts a base estimator instance
     - CARTs now use max features heuristic by default
+    - Added build/quick factory methods to Datasets
 
 - 0.0.5-alpha
     - Added Gaussian Mixture clusterer